PrisonersDilemmaPlayground.kt
Implemented the Prisoner's Dilemma experiment as its own class. Unit testing and integration testing remain to do, although manual testing suggests this early draft is working as expected. The goal is to create a more abstract Playground class of which PrisonersDilemmaPlayground is an instance, for tackling other problems -- but that's down the road.