Main.kt
For whatever underlying reason which deserves more research on my part, using runBlocking on the main function really hinders parallel computation of the entire program. Making main a suspending function which uses normal coroutineScope seems to allow for using much more of my PC's available processing power here. I am going to look in to this more, but for now I'm going with main() = coroutineScope {}, as that is much, much faster on my machine.