PrisonersDilemmaGame.kt
Switched to a "Resource-Based" system, as suggested by the book *Signals and Boundaries* (a 2012 successor to the 1992 paper which inspired this project). In this simple case, the resource is "wins" and a player can get 2 by scoring lower than an opponent or 1 by a tie. This leads very quickly to a pretty good pool of players. You can play with the threshold at which players have enough "wins" to reproduce in order to watch things play out at a different pace and there may be an optimal number for it (I'm not sure what that optimal number is yet though).