Balance Tweaks
The book *Signals and Boundaries* (by John Holland, the same person who wrote the *Genetic Algorithms* paper) suggests using a resource economy instead of a simple "average score", and taking that advice has been a good idea. It took a lot of tweaking to find the ideal mix to get it to do what I wanted, but the system is very flexible. Another thing suggested by that book is using Genetic Algorithms to tune other Genetic Algorithms. The parameters I've been tweaking could be auto-tuned by a pool of Playgrounds each trying to find the mixture of parameters which led to the optimal result most quickly -- a good project idea for the near future.